Closes #40.
Extracting muscle-to-force properties across arbitrary tendon-driven limbs.

Abstract:

Force control of a limb occurs under the patronage and limitation of neurological and musculoskeletal constraints. With elasticity of joints, inertia of bones, and dynamics of muscle excitation contributing to motor variability, how do animals translate tendon forces effectively into endpoint forces? In this study, we present a powerful method and an illustrative dataset for extracting posture-specific force control parameters from arbitrary limbs. We used an anatomic robotic finger, had it produce isometric fingertip forces upon a 6D force transducer, and used a stiff 6DOF robotic manipulator to move the endpoint to 1000 different fingertip postures. The resulting dataset features how 7 tendons produce 6 dimensions of endpoint force at a sampling rate of 1KHz—amounting to 98,125,444 observations of tension-to-force for muscles with up to 20N of tension. We discuss (i) detailed statistics of the variability in static muscle force, (ii) the comparison of linear model performance across a line of 1000 different postures, (iii) metrics for how many muscle activation patterns are needed to acquire a confident estimate of the feasible force space, (iv) a visualization of muscle-to-force coherence for each of the seven muscles, and (v) the technical approach for researchers to evaluate their muscle models across data of this composition and magnitude. Interestingly, we found that posture did not have a significant effect on linear model performance across different locations in endpoint force, with nearly all models reaching 70% of variance-accounted-for in output force. An application of feasibility theory to this work allows us to identify how force control strategies must change when a muscle is increased, or simulate how the feasibility space changes when muscles are paired together. This leads naturally to the quantification of muscle contribution to different force directions, at different postures, and informs how the afferent signals would change with muscle injury or fatigue. Ultimately, we discuss the implications of this method and dataset, with specific reference to evaluating muscle energetics (cost functions), muscle-force coherence (control policies), the effect of dynamics across squishy and noncompliant robotic and cadaveric limbs (plant dynamics), and how force-control data like these can contextualize the behaviors observed in-vivo.